sub(feat): don't build on npm install
The compatibility issue we are having is due to type declarations not
being backwards compatible with older version of TS. We need to manually
modify the declaration files to make them compatible. Because of this,
we shouldn't build on npm install because the backwards compatible type
declarations will be overwritten.